Finish css on homepage, make minor edits to pages
Set up basic css for homepage for background img, h1 and links to other
pages. Used background-attachment to fix background image so it would
not scale with the size of the viewport and leave gap under the image.

Also centered the recipe-img and home-link on recipe pages.

Definitely need to commit more often.